+++ doc/CHANGES.rst
@@ -22,6 +22,9 @@ CHANGELOG
   __emulates__ attributes. InstrumentedList haven't to have this attribute
 * [FIX] SQLAlchemy 1.1.*, Session State changed, update the update method 
   of the registry to install / update / uninstall bloks
+* [FIX] SQLAlchemy 1.1.*, Hybrid property don't propagate the relationship
+  info attribute. The propagate is forced for Many2One and One2One. The only
+  both relationships to be wrapped by hybrid_property
 
 0.9.2 (2016-10-12)
 ------------------